requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
numpy
scikit-learn
joblib
tensorflow-cpu
